New writer file begins
